Initial

+++ src/main/java/user/authentication/user_authentication/service/UserService.java
@@ -4,6 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.retry.annotation.Backoff;
 import org.springframework.retry.annotation.Retryable;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -12,6 +14,7 @@ import user.authentication.user_authentication.model.User;
 import user.authentication.user_authentication.repository.UserRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -34,10 +37,16 @@ public class UserService {
     }
 
     @Retryable(value = DataAccessException.class, maxAttempts = 3, backoff = @Backoff(delay = 1000))
-    public void deleteUser(String id) {
+    public ResponseEntity deleteUser(String id) {
         try {
             logger.info("Attempting to delete user with ID: {}", id);
-            userRepository.deleteById(id);
+            Optional<User> user = userRepository.findById(id);
+            if (user.isPresent()) {
+                userRepository.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (DataAccessException ex) {
             logger.error("Attempt to delete user failed", ex);
             throw ex;
